Add quick sort with list API for Homework2
In HW2, it requires us to implement quick sort with list API, so I make a branch in lab0 to
try to implement it under the rules of list.h and queue.h. The current branch using quick sort
can pass traces-03 but not traces-14 and 15, since it couldn't pass the time complexity test
due to the O(n^2) worse case rather than O(nlogn) of merge sort.