Delete queue.c for merging online
I don't know what the heck is teacher doing to master branch in origin
repository, I could not sync it to my fork and I don't quite know how to
resolve that in my own computer.

Hence, I want to try change the conflicting file to the way it may be
able to be merged.

Change-Id: I36404b2924a6ebe1decc041b2145f2e1a83e81de